# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MoldGuard Competition Server API
  version: 1.0.0
paths:
...
  /api/v1/work-orders:
    get:
      operationId: list_work_orders
      tags:
      - v1
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAPIEnvelope'
          description: ''
  /api/v1/work-orders/{work_order_id}:
    get:
      operationId: retrieve_work_order
      parameters:
      - in: path
        name: work_order_id
        schema:
          type: string
        required: true
      tags:
      - v1
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAPIEnvelope'
          description: ''
...
components:
  schemas:
...
    OpenAPIEnvelope:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          nullable: true
        errors: {}
        request_id:
          type: string
      required:
      - code
      - message
      - request_id
...